Benchmark improvements
DlibQueueWrapper, unlike every other wrappers, allows only
swap semantics, that is no `enqueue(T const &)`, only
`enqueue(T &)` or `enqueue(T &&)`. And that means we can't
send enqueue counter_t aka std::uint64_t to in queue. Not
sure if this is fixed as intended, but it compiles
and gives meaningful results.